check img type

+++ src/perception/src/CamNode.cpp
@@ -80,7 +80,7 @@ void ImgProcessing(const cv::Mat& img_frame, CAMERA_DATA* camera_data)
     KALMAN_STATE arst_KalmanState[2] = {0};
 
     // 이미지 전처리 진행할 임시 이미지 버퍼 생성 Camera.yaml의 remap size 참고
-    cv::Mat Temp_Img(camera_data->st_CameraParameter.s32_RemapHeight, camera_data->st_CameraParameter.s32_RemapWidth, CV_32FC1);
+    cv::Mat Temp_Img(camera_data->st_CameraParameter.s32_RemapHeight, camera_data->st_CameraParameter.s32_RemapWidth, CV_8UC1);
     cv::Mat st_NoneZero,st_Tmp, st_ResultImage(Temp_Img.size(), CV_8UC3, Scalar(0,0,0)); //결과랑 중간계산 mat, 최종시각화 검정색으로 초기화 
     // std::cout << img_frame.size() <<std::endl; // check origin img size (1280*720)
